+++ apps/www/src/registry/default/plate-ui/link-floating-toolbar.tsx
@@ -1,13 +1,17 @@
 import React from 'react';
-import { TEditableProps } from '@udecode/plate-common';
+import {
+  flip,
+  offset,
+  UseVirtualFloatingOptions,
+} from '@udecode/plate-floating';
 import {
   FloatingLinkUrlInput,
+  LinkFloatingToolbarState,
   LinkOpenButton,
   useFloatingLinkEdit,
   useFloatingLinkEditState,
   useFloatingLinkInsert,
   useFloatingLinkInsertState,
-  useFloatingLinkSelectors,
 } from '@udecode/plate-link';
 
 import { cn } from '@/lib/utils';
@@ -17,18 +21,45 @@ import { buttonVariants } from './button';
 import { inputVariants } from './input';
 import { popoverVariants } from './popover';
 import { Separator } from './separator';
+import { ToolbarProps } from './toolbar';
 
-export function LinkFloatingToolbar({ readOnly }: TEditableProps) {
-  const isEditing = useFloatingLinkSelectors().isEditing();
+const floatingOptions: UseVirtualFloatingOptions = {
+  placement: 'bottom-start',
+  middleware: [
+    offset(12),
+    flip({
+      padding: 12,
+      fallbackPlacements: ['bottom-end', 'top-start', 'top-end'],
+    }),
+  ],
+};
 
-  const state = useFloatingLinkInsertState();
+export interface LinkFloatingToolbarProps extends ToolbarProps {
+  state?: LinkFloatingToolbarState;
+}
+
+export function LinkFloatingToolbar({ state, children, ...props }) {
+  const insertState = useFloatingLinkInsertState({
+    ...state,
+    floatingOptions: {
+      ...floatingOptions,
+      ...state?.floatingOptions,
+    },
+  });
   const {
     props: insertProps,
     ref: insertRef,
+    hidden,
     textInputProps,
-  } = useFloatingLinkInsert(state);
+  } = useFloatingLinkInsert(insertState);
 
-  const editState = useFloatingLinkEditState();
+  const editState = useFloatingLinkEditState({
+    ...state,
+    floatingOptions: {
+      ...floatingOptions,
+      ...state?.floatingOptions,
+    },
+  });
   const {
     props: editProps,
     ref: editRef,
@@ -36,7 +67,7 @@ export function LinkFloatingToolbar({ readOnly }: TEditableProps) {
     unlinkButtonProps,
   } = useFloatingLinkEdit(editState);
 
-  if (readOnly) return null;
+  if (hidden) return null;
 
   const input = (
     <div className="flex w-[330px] flex-col">
@@ -66,7 +97,7 @@ export function LinkFloatingToolbar({ readOnly }: TEditableProps) {
     </div>
   );
 
-  const editContent = isEditing ? (
+  const editContent = editState.isEditing ? (
     input
   ) : (
     <div className="box-content flex h-9 items-center gap-1">