+++ packages/node-id/src/createNodeIdPlugin.ts
@@ -27,6 +27,12 @@ export interface NodeIdPlugin extends QueryNodeOptions {
    * @default false
    */
   reuseId?: boolean;
+
+  /**
+   * By default, when a node inserted using editor.insertNode(s) has an id, it will be used instead of the id generator, except if it already exists in the document.
+   * Set this option to true to disable this behavior.
+   */
+  disableInsertOverrides?: boolean;
 }
 
 export const KEY_NODE_ID = 'nodeId';