+++ packages/comments/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @udecode/plate-comments
 
+## 23.6.0
+
+## 23.3.1
+
+## 23.3.0
+
 ## 22.1.0
 
 ### Patch Changes